backend/requirements.txt
# Backend requirements for the FastAPI e‑commerce app
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
apscheduler
psycopg2-binary  # if using PostgreSQL; omit for SQLite
pydantic
pydantic[email]